Release 3.2.0: consolidated public API, extractor functions, and rewritten vignette suite
- Narrow the exported API to genuinely user-facing functions; five internal
  numerical-stability helpers and a low-level EM data-filling routine are
  no longer exported (still reachable via ::: for anyone relying on them).
- predict_lucid() and boot_lucid() auto-detect early/parallel/serial from
  the fitted model's class; lucid_model is now optional on both.
- Add four extractor functions that read directly off a fitted model and
  auto-detect its architecture: get_selected_G(), get_selected_Z(),
  get_cluster_assignment(), get_top_omics_features().
- summary()'s bootstrap CI tables (G-to-X, cluster-to-Y, and cluster-omics
  means) mark statistically significant rows with a sig column, based on
  the normal-theory interval; the percentile-interval columns are no
  longer shown in the printed table (still available on the boot_lucid()
  return object).
- Fix a bug where estimate_lucid()/tune_lucid() could not actually be
  called without explicitly naming lucid_model, contradicting their
  documented default.
- Fix get_selected_G() crashing when CoG covariates were supplied.
- Rewrite the vignette suite as polished, user-facing tutorials built
  entirely on the public API: lucid_3models_normal_outcome.Rmd and
  lucid_3models_binary_outcome.Rmd (early/parallel/serial walkthroughs,
  each fitting step shown explicitly), lucidus_full_functionality_guide.Rmd,
  lucid_early_sankey_top_omics_tutorial.Rmd, and
  serial_6stage_parallel_monotone_missing_tutorial.Rmd. Rendered copies are
  included under docs/ and linked from the README.
- Refresh README.Rmd/README.md with a self-contained example on the
  bundled HELIX dataset: load data, fit a parallel-integration model,
  read the summary, extract results, and visualize cluster omics profiles.
- Add NEWS.md documenting the above.
- Expand the test suite to 10,844 assertions covering the corrections and
  additions above; full suite passes with zero failures.

+++ tests/testthat/test-gcomp-binary-1-layer.R
@@ -32,7 +32,7 @@ test_that("check variable selection on G", {
 
   # compare prediction of X
   expect_equal(fit1$inclusion.p, pred1$inclusion.p, tolerance = 0.05)
-  expect_equal(class(pred1$pred.x), "integer")
+  expect_equal(class(pred1$pred.x), "numeric")
 
   pred2 <- predict_lucid(model = fit1,
                          lucid_model = "early",
@@ -41,7 +41,7 @@ test_that("check variable selection on G", {
                          Y = NULL,
                          CoY = cov)
 
-  expect_equal(class(pred2$pred.x), "integer")
+  expect_equal(class(pred2$pred.x), "numeric")
   
 
   counter_G_v1_1 <- G
